requirements.txt
cassandra-driver==3.23.0